Moved function to Arrays.php
